backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pgvector==0.2.4
cors==1.0.1
fastapi-cors==0.0.6
